added a solution for dealing with 15 fps that just doubles frames added to buffer. added visualization of read and write positions. closes #8.

+++ AndyWarholMachine/src/testApp.h
@@ -32,7 +32,7 @@ public:
 	
 	Grapher avgGrapher, devGrapher;
 	
-	int cameraFrames;
+	unsigned long cameraFrames;
 	
 	ofxAutoControlPanel panel;
